Implement suggested improvements
Boris Dalstein, the author of VCC made some suggestions to improve the code.  This is the first of those implementations.

Boris wrote:  
```you've used inputtransform to transform the positions. This is incorrect: inputtransform is only meant for transforming the inputpositions, if any. You should just use the positions as is. In the future, there will be a transform attribute for the purpose that you thought inputtransform was. The inputpositions are a record of the raw mouse/tablet samples provided by the OS, in window coordinates, and is stored in the file only if "Experimental > Save Input Sketch Points" is checked. Then inputtransform is a record of the current canvas position/scale/orientation when you drew that specific stroke, and it is necessary to store it for "Experimental > Show Input Sketch Points" to work. Currently, I think your exporter only works if you don't move the canvas at all while drawing your whole illustration :)```

Signed-off-by: Rodney <rodney.baker@gmail.com>